requirements.txt
streamlit==1.31.1
pandas==2.1.4
gspread==5.12.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0